Shape up pipeline
